requirements.txt
aniso8601==7.0.0
attrs==19.1.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Click==7.0
Flask==1.1.1
flask-accepts==0.10.0
Flask-RESTful==0.3.7
flask-restplus==0.12.1
flask-restx==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
marshmallow==3.2.0
mypy-extensions==0.4.1
packaging==20.9
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.15.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2019.2
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.6
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==0.15.5
